Update for compatibility with modern rust versions.

+++ obs/src/lib.rs
@@ -21,7 +21,7 @@ use std::{
 
 pub unsafe fn register_source(info: &'static sys::obs_source_info, info_size: Option<usize>) {
     let info_size = info_size.unwrap_or(mem::size_of::<sys::obs_source_info>());
-    sys::obs_register_source_s(info as *const _, info_size as u64);
+    sys::obs_register_source_s(info as *const _, info_size as usize);
 }
 
 #[macro_export]
